update task bug fixed

+++ app/Services/Task/UpdateTaskService.php
@@ -282,17 +282,22 @@ class UpdateTaskService
     private function updateTask($task, $user): void
     {
         $cacheValues = cache()->get('task_update_' . $task->id);
-        // Save task custom fields
-        $task->custom_field_values()->delete();
-        foreach ($cacheValues['custom_fields'] as $customField) {
-            $customField['task_id'] = $task->id;
-            CustomFieldsValue::query()->create($customField);
+        if (array_key_exists('custom_fields', $cacheValues)) {
+            // Save task custom fields
+            $task->custom_field_values()->delete();
+            foreach ($cacheValues['custom_fields'] as $customField) {
+                $customField['task_id'] = $task->id;
+                CustomFieldsValue::query()->create($customField);
+            }
         }
-
-        // Save task addresses
-        $task->addresses()->delete();
-        foreach ($cacheValues['addresses'] as $address) {
-            Address::query()->create($address);
+        $addressesCount = 0;
+        if (array_key_exists('addresses', $cacheValues)) {
+            // Save task addresses
+            $addressesCount = count($cacheValues['addresses']);
+            $task->addresses()->delete();
+            foreach ($cacheValues['addresses'] as $address) {
+                Address::query()->create($address);
+            }
         }
 
         $task->update([
@@ -308,7 +313,8 @@ class UpdateTaskService
             'status' => Task::STATUS_OPEN,
             'user_id' => $user->id,
             'phone' => $user->phone_number,
-            'coordinates' => $cacheValues['addresses'][0]['latitude'] . ',' . $cacheValues['addresses'][0]['longitude']
+            'coordinates' => $addressesCount > 0 ? $cacheValues['addresses'][0]['latitude'] . ',' . $cacheValues['addresses'][0]['longitude'] : ''
         ]);
+
     }
 }
